new button, no scroll, and spacebar...
There's two buttons, the start button navigates to the menu scene and the back button stays on the scene and makes both buttons disappear. These buttons are stuck to the screen, they do not move. Also, when you press the spacebar, all input for the car is ignored. I assigned this function to the spacebar to test out when the car hits the checkpoint and a text box appears and the car stops. TAGGART ADD THE HOWTO TILE MAP!!!!!